requirements.txt
pip>=9.0.1
urllib3[secure]
web.py
requests
cffi
cryptography
msrest
msrestazure
adal
azure-common
azure-keyvault
